+++ apps/web/src/app/api/comments/route.ts
@@ -1,8 +1,8 @@
 import CommentService from '@/services/comment.service';
 import { NextRequest, NextResponse } from 'next/server';
-import { verifyRequestCookies } from '@/utils/cookie.server';
+import { verifyRequestCookies } from '@/shared/utils/cookie.server';
 import { ResponseError } from '@moneed/utils';
-import { ERROR_MSG, SUCCESS_MSG } from '@/constants/message';
+import { ERROR_MSG, SUCCESS_MSG } from '@/shared/config/message';
 
 export async function POST(req: NextRequest) {
     try {
