+++ apps/web/src/app/api/posts/route.ts
@@ -1,9 +1,9 @@
-import { verifyRequestCookies, assertAccessTokenPayload } from '@/utils/cookie.server';
+import { verifyRequestCookies, assertAccessTokenPayload } from '@/shared/utils/cookie.server';
 import PostService from '@/services/post.service';
 import { StockService } from '@/features/stock/service/stock.server';
 import { NextRequest, NextResponse } from 'next/server';
 import { ResponseError } from '@moneed/utils';
-import { ERROR_MSG, SUCCESS_MSG } from '@/constants/message';
+import { ERROR_MSG, SUCCESS_MSG } from '@/shared/config/message';
 
 export async function GET(req: NextRequest) {
     try {
