+++ apps/web/src/app/api/users/nickname/check/route.ts
@@ -1,8 +1,8 @@
 import { NextResponse } from 'next/server';
 import UserService from '@/services/user.service';
-import { verifyRequestCookies, assertAccessTokenPayload } from '@/utils/cookie.server';
+import { verifyRequestCookies, assertAccessTokenPayload } from '@/shared/utils/cookie.server';
 import { ResponseError } from '@moneed/utils';
-import { ERROR_MSG, SUCCESS_MSG } from '@/constants/message';
+import { ERROR_MSG, SUCCESS_MSG } from '@/shared/config/message';
 
 export async function POST(request: Request) {
     try {
