+++ apps/web/src/app/myprofile/page.tsx
@@ -9,7 +9,7 @@ import { checkDuplicateNickname, updateUserProfile } from '@/apis/user.api';
 import { useMutation } from '@tanstack/react-query';
 import { useForm } from 'react-hook-form';
 import { AxiosError } from 'axios';
-import { REASON_CODES } from '@/constants/snackbar';
+import { REASON_CODES } from '@/shared/config/snackbar';
 import { MyProfileSkeleton } from '@/components/Skeletons/MyProfileSkeleton';
 import { useDebounce } from '@/shared/hooks/useDebounce';
 import { useModal } from '@/context/ModalContext';
