+++ apps/web/src/app/selectstocktype/page.tsx
@@ -1,1 +1,1 @@
-export { SelectStockPage as default } from '@/_pages/select-stock';
+export { default } from '@/_pages/select-stock';
