+++ apps/web/src/services/short.service.ts
@@ -1,5 +1,5 @@
 import { searchShorts } from '@/apis/youtube.api';
-import { ERROR_MSG } from '@/constants/message';
+import { ERROR_MSG } from '@/shared/config/message';
 import type { Shorts } from '@prisma/client';
 import ShortRepository from '@/repositories/short.repository';
 import { parseShorts } from '@/utils/parser';
