+++ apps/web/src/shared/api/client.ts
@@ -1,10 +1,10 @@
 // import { REASON_CODES } from '@/constants/snackbar';
-import { ERROR_MSG } from '@/constants/message';
+import { ERROR_MSG } from '@/shared/config/message';
 import axios from 'axios';
 import { AxiosInstance, AxiosError, InternalAxiosRequestConfig, AxiosResponse } from 'axios';
 import { refresh } from '../../apis/auth.api';
 import { getAccessToken, getRefreshToken, clearTokens, setTokens } from '@/utils/localStorage.browser';
-import { REASON_CODES } from '@/constants/snackbar';
+import { REASON_CODES } from '@/shared/config/snackbar';
 
 const moneedBaseConfig = {
     withCredentials: true,
