+++ apps/web/src/shared/utils/cookie.browser.ts
