+++ apps/web/src/shared/utils/server.ts
@@ -1,0 +1,1 @@
+export { verifyRequestCookies } from './cookie.server';
